+++ mailing/services.py
@@ -1,6 +1,21 @@
 from crontab import CronTab
 from django.core.mail import send_mail
 from django.conf import settings
+from django.template.defaultfilters import slugify as d_slugify
+
+
+def convert_word(words) -> str:
+    """
+    Slugify for russian language.
+    """
+    alphabet = {'а': 'a', 'б': 'b', 'в': 'v', 'г': 'g', 'д': 'd', 'е': 'e', 'ё': 'yo', 'ж': 'zh', 'з': 'z',
+                'и': 'i',
+                'й': 'j', 'к': 'k', 'л': 'l', 'м': 'm', 'н': 'n', 'о': 'o', 'п': 'p', 'р': 'r', 'с': 's', 'т': 't',
+                'у': 'u', 'ф': 'f', 'х': 'kh', 'ц': 'ts', 'ч': 'ch', 'ш': 'sh', 'щ': 'shch', 'ы': 'i', 'э': 'e',
+                'ю': 'yu',
+                'я': 'ya'}
+
+    return d_slugify(''.join(alphabet.get(w, w) for w in words.lower()))
 
 
 def sendmail(to, subject, message):
@@ -17,4 +32,3 @@ def set_cron():
         job = cron.new(command='echo hello_world')
         job.minute.every(1)
     print('cron.write() was just executed')
-
