+++ mailing/templates/mailing/main.html
@@ -1,4 +1,5 @@
 {% extends 'mailing/base.html' %}
+{% load mailing_tags %}
 {% block body %}
     {% if Blog %}
     <div class="pricing-header p-3 pb-md-4 mx-auto text-center">
@@ -14,7 +15,9 @@
           </div>
             <div class="card-body">
              <ul class="list-unstyled mt-3 mb-4">
-              <li>{{ item.plot|slice:50 }} </li>
+               {% if item.image%}
+                    <li><img src="{% media_path item.image %}" width="150""></li>
+               {% endif %}
               </ul>
                 <a href="{% url 'blog:blog_card' item.slug %}" class="btn btn-sm btn-outline-secondary">View</a>
             </div>
