--- mailing/utils.py
@@ -1,15 +1,0 @@
-from django.template.defaultfilters import slugify as d_slugify
-
-
-def convert_word(words) -> str:
-    """
-    Slugify for russian language.
-    """
-    alphabet = {'а': 'a', 'б': 'b', 'в': 'v', 'г': 'g', 'д': 'd', 'е': 'e', 'ё': 'yo', 'ж': 'zh', 'з': 'z',
-                'и': 'i',
-                'й': 'j', 'к': 'k', 'л': 'l', 'м': 'm', 'н': 'n', 'о': 'o', 'п': 'p', 'р': 'r', 'с': 's', 'т': 't',
-                'у': 'u', 'ф': 'f', 'х': 'kh', 'ц': 'ts', 'ч': 'ch', 'ш': 'sh', 'щ': 'shch', 'ы': 'i', 'э': 'e',
-                'ю': 'yu',
-                'я': 'ya'}
-
-    return d_slugify(''.join(alphabet.get(w, w) for w in words.lower()))
